feat (user views)

+++ resources/views/livewire/admin/products/index.blade.php
@@ -1,7 +1,7 @@
 <div class="font-family p-2">
     <div class="flex justify-between items-center mb-2">
         <h1 class="text-2xl font-bold text-blue-700 mb-3">Produtos</h1>
-        <a href="{{ route('admin.products.create')}}"><button class="text-sm bg-blue-600 border border-blue-600 hover:bg-blue-700 text-white py-1 px-3 rounded-md">Novo Produto</button></a>
+        @if (auth()->user()->role == "Dono") <a href="{{ route('admin.products.create')}}"><button class="text-sm bg-blue-600 border border-blue-600 hover:bg-blue-700 text-white py-1 px-3 rounded-md">Novo Produto</button></a> @endif
     </div>
     <div class="relative overflow-x-auto rounded-md">
         <input wire:model.lazy="search" type="text" placeholder="Pesquise aqui" name="search" id="search" class="px-3 rounded-md border-0 py-1.5 mb-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 sm:text-sm sm:leading-6">
@@ -20,9 +20,9 @@
                     <th scope="col" class="px-6 py-3">
                         Preço
                     </th>
-                    <th scope="col" class="px-6 py-3">
+                    @if (auth()->user()->role == "Dono")  <th scope="col" class="px-6 py-3">
                         Valor Pago
-                    </th>
+                    </th> @endif
                     <th scope="col" class="px-6 py-3">
                         IMEI
                     </th>
@@ -50,16 +50,16 @@
                     <td class="px-6 py-4">
                         R${{ $product->price  }}
                     </td>
-                    <td class="px-6 py-4">
+                    @if (auth()->user()->role == "Dono") <td class="px-6 py-4">
                         R${{ $product->bought_value  }}
-                    </td>
+                    </td> @endif
                     <td class="px-6 py-4">
                         {{ $product->imei  }}
                     </td>
                     <td class="px-6 py-4">
                         {{ $product->created_at->format('d/m/Y')  }}
                     </td>
-                    <td class="px-6 py-4">
+                    @if (auth()->user()->role == "Dono")  <td class="px-6 py-4">
                       <button wire:click="deleteProduct({{ $product->id }})">
                           <i class="fa fa-trash text-red-500"></i>
                       </button>
@@ -71,7 +71,7 @@
                       </button></a>
                       <a href="{{  route('admin.products.photos', $product->id) }}"><button class="">
                           <i class="fa-solid fa-file-image text-indigo-500"></i>
-                      </button></a>
+                      </button></a> @endif
                       <a href="{{  route('sales.create', $product->id) }}"><button class="">
                           <i class="fa fa-wallet text-yellow-500"></i>
                       </button></a>
